+++ qjazz-processes/qjazz_processes/processing/inputs/layers.py
@@ -164,7 +164,6 @@ class ParameterVectorLayer(ParameterMapLayer):
 
 
 class ParameterFeatureSource(ParameterMapLayer):
-
     _SchemaFormat = "x-feature-source"
 
     @classmethod