+++ qjazz-processes/qjazz_processes/schemas/bbox.py
@@ -11,7 +11,6 @@ from typing import (
 from annotated_types import Len
 from pydantic import Field
 
-from .crs import CrsDefinition
 from .models import JsonModel, OneOf
 from .ogc import WGS84
 
@@ -24,10 +23,8 @@ BboxCoordinates = OneOf[Extent2D | Extent3D]
 # OGC bounding box definition
 # See https://schemas.opengis.net/ogcapi/processes/part1/1.0/openapi/schemas/bbox.yaml
 def BoundingBox(default: Optional[str] = None) -> TypeAlias:
-
     class BBox(JsonModel):
         bbox: BboxCoordinates
-        crs: CrsDefinition = Field(default or WGS84)  # type: ignore [valid-type]
+        crs: str = Field(default or WGS84)  # type: ignore [valid-type]
 
     return BBox
-